Change the WebSql database size to prevent permission dialog on Safari
Due to a slightly different interpretation of w3.org's recommended
database space limit of "five megabytes" Safari will open a dialog
asking "Allow this website to use space on your disk?" if the size
requested is greater than 5*1000*1000 bytes, whereas Kinvey always
requests 5*1024*1024 bytes. The definition of megabyte is unfortunately
ambiguous so there is no correct interpretation. Since the WebSqlAdapter
size was presumably chosen to avoid this dialog, we should set it to the
more conservative of the two interpretations.